Housekeep command-based testing harness
