+++ signer/tests/integration/commands/context.rs
@@ -1,24 +1,31 @@
+use std::sync::Arc;
+
 use madhouse::{State, TestContext};
+use signer::{
+    storage::{model::EncryptedDkgShares, postgres::PgStore},
+    testing::storage::new_test_database,
+};
+use tokio::runtime::Runtime;
 
 #[derive(Clone, Debug)]
 pub struct Ctx {
-    db: signer::storage::postgres::PgStore,
-    runtime: std::sync::Arc<tokio::runtime::Runtime>,
+    db: PgStore,
+    runtime: Arc<Runtime>,
 }
 
 impl Ctx {
     pub fn new() -> Self {
-        let runtime = std::sync::Arc::new(tokio::runtime::Runtime::new().unwrap());
-        let db = runtime.block_on(async { signer::testing::storage::new_test_database().await });
+        let runtime = Arc::new(Runtime::new().unwrap());
+        let db = runtime.block_on(async { new_test_database().await });
 
         Self { db, runtime }
     }
 
-    pub fn db(&self) -> &signer::storage::postgres::PgStore {
+    pub fn db(&self) -> &PgStore {
         &self.db
     }
 
-    pub fn runtime_handle(&self) -> std::sync::Arc<tokio::runtime::Runtime> {
+    pub fn runtime_handle(&self) -> Arc<Runtime> {
         self.runtime.clone()
     }
 }
@@ -27,7 +34,7 @@ impl TestContext for Ctx {}
 
 #[derive(Debug, Default)]
 pub struct TestState {
-    pub shares: Option<signer::storage::model::EncryptedDkgShares>,
+    pub shares: Option<EncryptedDkgShares>,
 }
 
 impl State for TestState {}
