+++ signer/tests/integration/commands/mod.rs
@@ -1,9 +1,7 @@
 mod context;
-mod db;
 mod dkg;
 mod test_cases;
 
 pub use context::{Ctx, TestState};
-pub use db::WriteDkgShares;
-pub use dkg::CreateFailedDkgShares;
+pub use dkg::{CreateFailedDkgShares, WriteDkgShares};
 pub use test_cases::VerifyDkgVerificationFailed;
